Emit `#nowarn` for obsolete usage & invalid XML comments
Fixes #405: `#nowarn` for obsolete usage
* Usage is quite hard to detect -> always emit when there's at least one
`Obsolete` (`@deprected` in Typescript)

Fixes #409: `#nowarn` for invalid xml comments
* Invalid xml comments aren't easy to detect (invalid namings,
difference in paths for references, ...) -> always emit when there's a
least on XML comment with tags

Add `AdditionalData` on `FsFileOut`: Text that is printed at specified
location in output file.
Used for `#nowarn`, but is quite useful for logging output too.