Update fable to `3.3.1`
Note: There are some changes in `3.4` which result in an error:
```
[...]\ts2fable\build\test\functionTests.js:1
Error [ERR_REQUIRE_ESM]: require() of ES Module [...]\ts2fable\build\test\fable_modules\fable-library.3.4.0\Util.js not supported.
Instead change the require of Util.js in null to a dynamic import() which is available in all CommonJS modules.
```
-> not yet updated to `3.4.x`